backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bidict==0.23.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
ecdsa==0.19.1
fastapi==0.115.12
greenlet==3.2.0
h11==0.14.0
idna==3.10
mypy-extensions==1.0.0
packaging==24.2
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.7
psycopg2-binary==2.9.10
pyasn1==0.4.8
pydantic==2.11.3
pydantic_core==2.33.1
python-dotenv==1.1.0
python-engineio==4.12.0
python-jose==3.4.0
python-multipart==0.0.20
python-socketio==5.13.0
requests==2.32.3
rsa==4.9.1
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.1
wsproto==1.2.0
